Correct wrong signature for Object.wait(JI)V.

+++ runt/apis/cldc-compact/cc/squirreljme/runtime/cldc/asm/ObjectAccess.java
@@ -138,7 +138,7 @@ public final class ObjectAccess
 	 * @since 2018/11/21
 	 */
 	public static final native int monitorWait(Object __o, long __ms,
-		long __ns);
+		int __ns);
 	
 	/**
 	 * Constructs and initializes a new instance of the class by the given
